Update README.md
After clicking Support Sam by giving him money via Patreon, it showed 404 so I change the link from https://patreon.com/sammakesvr to https://www.patreon.com/SamMakesThings. The link can direct to the correct site, Samperson and select a membership level - heroes, legends and gods.

 After clicking How To Stop Desktop Goose, it showed 404 too so I change the link https://github.com/DesktopGooseUnoffical/ResourceHub/blob/master/tutorials/StopDesktopGoose.md to https://www.youtube.com/watch?v=pIpmH964mUk. The link direct to a video how to delete desktop goose, first the person keep clicking the program and there are 10 gooses on his desktop then press ESC button to delete goose.

For Q&A section, the question and answer was in the same sentence so I separate the answer from the question by beginning the answer on a new line.